Abstract
Abstract - Sensors with electrothermal excitation 

Silicon bars (thickness: approx. 50 μm, width: 1 mm) of different lengths (3 – 10 mm) were produced by anisotropic etching technology in thermally oxidized (100)-Si wafers and equipped with NiCr micro-heaters positioned at the bar ends. On the beam there are also four measuring resistors connected to form a full bridge. As a result of a pulse-like heating, the beam is excited to natural oscillations. The periodic expansion or compression of the measuring resistors results in a frequency-analogue bridge signal which is very well suited for detecting changes in resonance behaviour. The possibilities of this sensor principle for the detection of force, temperature and gas flow are shown by measurements. The realization of frequency-analog sensors has a cost advantage in that the measurement signal is quasi-digital, so that an A/D conversion is not necessary. The electrical layout, including the bondable and/or solderable external connections, was implemented using metal thin-film technology throughout. The layer systems NiCr / TiPdAu (solder version) and NiCr / TiAu / galvanic Au (bond version) were deposited by PVD and subjected to the usual photolithographic and wet-chemical process steps for structuring.